split pending invitations controller into two based on scopes
